refactored seed data script and created seed json files
