+++ script/seedData/itemPhotoSeedData.js
@@ -1,0 +1,56 @@
+'use strict'
+
+const itemPhotos = [
+  {
+    photoTitle: 'default.jpeg',
+    downloadURL:
+      'https://images-na.ssl-images-amazon.com/images/I/81vkislowDL._AC_SL1500_.jpg',
+    itemId: 9,
+  },
+  {
+    photoTitle: 'default.jpeg',
+    downloadURL: 'public/images/freeShareDefaultPic.jpeg',
+    itemId: 8,
+  },
+  {
+    photoTitle: 'default.jpeg',
+    downloadURL: 'public/images/freeShareDefaultPic.jpeg',
+    itemId: 7,
+  },
+  {
+    photoTitle: 'default.jpeg',
+    downloadURL: 'public/images/freeShareDefaultPic.jpeg',
+    itemId: 3,
+  },
+  {
+    photoTitle: 'boots',
+    cloudREF: '/images/boots.webp',
+    downloadURL:
+      'https://firebasestorage.googleapis.com/v0/b/freeshare-7b345.appspot.com/o/images%2Fboots.webp?alt=media&token=4cc149ba-2e57-4397-8a89-3d5c79a07c99',
+    itemId: 5,
+  },
+  {
+    photoTitle: 'pan',
+    cloudREF: '/images/pan.jfif',
+    downloadURL:
+      'https://firebasestorage.googleapis.com/v0/b/freeshare-7b345.appspot.com/o/images%2Fpan.jfif?alt=media&token=75c2ea85-57ac-48c2-af08-69f83b9f9f0b',
+    itemId: 2,
+  },
+  {
+    photoTitle: 'stove',
+    cloudREF: '/images/stove.jpg',
+    downloadURL:
+      'https://firebasestorage.googleapis.com/v0/b/freeshare-7b345.appspot.com/o/images%2Fpan.jfif?alt=media&token=75c2ea85-57ac-48c2-af08-69f83b9f9f0b',
+
+    itemId: 2,
+  },
+  {
+    photoTitle: 'watch',
+    cloudREF: '/images/watch.jpg',
+    downloadURL:
+      'https://firebasestorage.googleapis.com/v0/b/freeshare-7b345.appspot.com/o/images%2Fwatch.jpg?alt=media&token=cb3cfdb2-bcda-44de-a289-c6eb1b8abf82',
+    itemId: 4,
+  },
+]
+
+module.exports = itemPhotos
