+++ script/seedData/itemSeedData.js
@@ -1,0 +1,90 @@
+'use strict'
+
+//item data
+const items = [
+  {
+    itemListName:
+      'IBIS Titanium Racing Road Bike USA SEBASTOPOL 56cm Made in California',
+    description:
+      'Ibis TI Bike, highly collectible, Superb titanium tube set, superb ride, 56CM top tube, semi sloping geo. 52cm seat tube. FYI, I am keeping the saddle, wheel set and pedals for my other bike.Questions welcomed.',
+    itemType: 'Offer',
+    itemCondition: 'Used',
+    status: 'Closed',
+    deliveryOption: 'Pickup_Only',
+    userId: 1,
+  },
+  {
+    itemListName: 'stove and pan',
+    description:
+      'Another band member left this in my flight case 15 years ago and does not want it back. I don’t have microphones or an AC adapter for it. I have never even tried to plug it in or turn it on.If you want it, please email me with a possible pickup time.',
+    itemType: 'Offer',
+    itemCondition: 'Used',
+    status: 'Open',
+    deliveryOption: 'Pickup_Only',
+    userId: 1,
+  },
+  {
+    itemListName: 'Queen bed frame and duvet',
+    description:
+      "A queen metal bedframe. Also duvet + cover + pillow. No mattress. The duvet has only been lightly used and are only several months old / the duvet cover just needs a regular wash in the laundry. I would have kept them but I needed to move fast and wasn't able to take them with me.",
+    itemType: 'Offer',
+    itemCondition: 'Used',
+    status: 'Open',
+    deliveryOption: 'Pickup_Only',
+    userId: 2,
+  },
+  {
+    itemListName: 'Gucci Gold Sandals',
+    description:
+      'Gold Sandals in good condition. It comes with the box and a dust bag. Worn once. Size 38.',
+    itemType: 'Offer',
+    itemCondition: 'Gently_Used',
+    status: 'Open',
+    deliveryOption: 'Pickup_Only',
+    userId: 4,
+  },
+  {
+    itemListName: 'Gucci Boots',
+    description:
+      'ISO of Gucci Boots since I lost my pair of Gucci sandals. Size 38.',
+    itemType: 'Seeking',
+    itemCondition: 'New',
+    status: 'Open',
+    userId: 4,
+  },
+  {
+    itemListName: 'FREE 87 REGAL AFT CABIN-NEEDS MOTOR',
+    description:
+      '87 REGAL 23 FT. NEEDS MOTOR. FULL V CABIN PLUS AFT CABIN! SLEEPS 6!! ,FRIDGE SINK,STOVE,FULL HEAD WITH TANK! JUST BOTTOM PAINTED 5 YR PAINT! Never got around to putting in an engine and was just going to donate to make room for my other boat. hard to work with the charitable org. during this pandemic so I figured I give it a shot to sell for same as donation. If you Know way around boats could pick up a deal. Boat is in water and would have to pick it up at nearby ramp.Very LOCAL Water delivery also possible. Thanks.',
+    itemType: 'Offer',
+    itemCondition: 'Used',
+    status: 'Open',
+    userId: 3,
+  },
+  {
+    itemListName: 'Bike',
+    description: 'ISO of Bike',
+    itemType: 'Seeking',
+    itemCondition: 'Used',
+    status: 'Open',
+    userId: 3,
+  },
+  {
+    itemListName: 'Bike Helmet',
+    description: 'ISO of Bike Helmet',
+    itemType: 'Seeking',
+    itemCondition: 'New',
+    status: 'Open',
+    userId: 3,
+  },
+  {
+    itemListName: 'A puppy',
+    description: 'ISO of cute puppy - prefer a pug',
+    itemType: 'Seeking',
+    itemCondition: 'New',
+    status: 'Closed',
+    userId: 1,
+  },
+]
+
+module.exports = items
