+++ script/seedData/seedDataIndex.js
@@ -1,0 +1,9 @@
+const users = require('./userSeedData')
+const items = require('./itemSeedData')
+const itemPhotos = require('./itemPhotoSeedData')
+
+module.exports = {
+  users,
+  items,
+  itemPhotos,
+}
