+++ script/seedData/userSeedData.js
@@ -1,0 +1,68 @@
+'use strict'
+
+//userData
+const users = [
+  {
+    firstName: `Diego`,
+    middleName: ``,
+    lastName: `Abreu`,
+    email: `diego@mail.com`,
+    admin: true,
+    password: `password`,
+    street1: `3328 Tiemann Ave`,
+    street2: ``,
+    city: `Bronx`,
+    state: 'NY',
+    zip: `10469`,
+    latitude: 40.87645,
+    longitude: -73.84273,
+  },
+
+  {
+    firstName: `Melinda`,
+    middleName: ``,
+    lastName: `Armbruster`,
+    email: `melinda@mail.com`,
+    password: `password`,
+    street1: `31 Saint Johns Pl`,
+    street2: ``,
+    city: `Brooklyn`,
+    state: 'NY',
+    zip: `11217`,
+    latitude: 40.67824,
+    longitude: -73.97802,
+  },
+
+  {
+    firstName: `Yoshie`,
+    middleName: ``,
+    lastName: `Fujiwara`,
+    email: `yoshie@mail.com`,
+    password: `password`,
+    street1: `115 Broadway`,
+    street2: `3A`,
+    city: `New York`,
+    state: 'NY',
+    zip: `10006`,
+    latitude: 40.708832,
+    longitude: -74.011497,
+  },
+
+  {
+    firstName: `Jae`,
+    middleName: ``,
+    lastName: `Chung`,
+    email: `Jae@mail.com`,
+    admin: true,
+    password: `password`,
+    street1: `222 E 93rd St`,
+    street2: `#APT 11C`,
+    city: `New York`,
+    state: 'NY',
+    zip: `10002`,
+    latitude: 40.78231,
+    longitude: -73.94947,
+  },
+]
+
+module.exports = users
